_demo/requirements.txt
flask>=3.0
pyyaml>=6.0
python-frontmatter>=1.1
python-dateutil>=2.9
